Switch to a retry and fallback with feature flag
If the current authentication user is "dave" (or
configure that at app.services.user) attempt to get to the
backend. Otherwise, or if the attempt fails, use a fallback.

The existence of the fallback acknowledges that the UI/gateway
has some knowledge of the existence of the backend. The fallback
even has an (empty) angular module that would be provided by the
backend if it were there.